requirements.txt
sacrebleu
unbabel-comet
transformers
datasets